Fix an issue where resolveKill() emits a warning if the future was never started
Summary:
Ref T2794. This is a minor warning I hit while exploring Phage stuff. In Phage, we can `resolveKill()` a future that we never called `start()` on, which tries to `proc_terminate()` a process which never spawned. This emits a warning.

Instead, check that `$this->proc` exists before signaling it.

Test Plan: Added a unit test.

Reviewers: chad

Reviewed By: chad

Maniphest Tasks: T2794

Differential Revision: https://secure.phabricator.com/D17354

+++ src/future/exec/ExecFuture.php
@@ -426,7 +426,10 @@ final class ExecFuture extends PhutilExecutableFuture {
   public function resolveKill() {
     if (!$this->result) {
       $signal = 9;
-      proc_terminate($this->proc, $signal);
+
+      if ($this->proc) {
+        proc_terminate($this->proc, $signal);
+      }
 
       $this->result = array(
         128 + $signal,
